CB#7 - Refactoring - Remove auth feature logic from Infrastructure project
