+++ src/crypto-bank.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 using crypto_bank.Infrastructure.Common;
-using crypto_bank.Infrastructure.Features.Auth;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -11,8 +10,6 @@ public static class InfrastructureServiceCollectionExtensions
         this IServiceCollection services,
         ConfigurationManager configuration)
     {
-        services.AddAuth(configuration);
-
         services.AddSingleton<IPasswordHasher, PasswordHasher>();
 
         return services;
